Run event pipeline on worker threads
